# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EDCOCR API",
    "description": "REST API for forensic-grade document OCR processing.",
    "version": "4.1.0"
  },
  "paths": {
...
    "/api/v1/entities": {
      "get": {
        "tags": [
          "recall"
        ],
        "summary": "Search Entities",
        "description": "Search indexed entities across all processed documents.\n\nSupports filtering by entity type, text content, job ID, and\nminimum confidence. Results are ordered by most recently indexed.",
        "operationId": "search_entities_api_v1_entities_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity type (e.g. PERSON, DATE)",
              "title": "Type"
            },
            "description": "Filter by entity type (e.g. PERSON, DATE)"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text search query (LIKE matching)",
              "title": "Q"
            },
            "description": "Text search query (LIKE matching)"
          },
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job ID",
              "title": "Job Id"
            },
            "description": "Filter by job ID"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Minimum confidence threshold",
              "default": 0.0,
              "title": "Min Confidence"
            },
            "description": "Minimum confidence threshold"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/extractions": {
      "get": {
        "tags": [
          "recall"
        ],
        "summary": "Search Extractions",
        "description": "Search indexed key-value extractions across all processed documents.\n\nSupports filtering by field name, value content, job ID, and\nminimum confidence. Results are ordered by most recently indexed.",
        "operationId": "search_extractions_api_v1_extractions_get",
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by field name (e.g. invoice_number, date)",
              "title": "Field"
            },
            "description": "Filter by field name (e.g. invoice_number, date)"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Value search query (LIKE matching)",
              "title": "Q"
            },
            "description": "Value search query (LIKE matching)"
          },
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job ID",
              "title": "Job Id"
            },
            "description": "Filter by job ID"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Minimum confidence threshold",
              "default": 0.0,
              "title": "Min Confidence"
            },
            "description": "Minimum confidence threshold"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntitySearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/EntitySearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total",
          "limit",
          "offset"
        ],
        "title": "EntitySearchResponse",
        "description": "Paginated entity search results."
      },
      "EntitySearchResult": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "job_id",
          "entity_type",
          "text",
          "confidence",
          "source",
          "page",
          "document_name"
        ],
        "title": "EntitySearchResult",
        "description": "Single entity search result."
      },
...
      "ExtractionSearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ExtractionSearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total",
          "limit",
          "offset"
        ],
        "title": "ExtractionSearchResponse",
        "description": "Paginated extraction search results."
      },
      "ExtractionSearchResult": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "field_name": {
            "type": "string",
            "title": "Field Name"
          },
          "field_value": {
            "type": "string",
            "title": "Field Value"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name"
          }
        },
        "type": "object",
        "required": [
          "extraction_id",
          "job_id",
          "field_name",
          "field_value",
          "confidence",
          "page",
          "document_name"
        ],
        "title": "ExtractionSearchResult",
        "description": "Single extraction search result."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Set via OCR_API_KEY environment variable."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth2 Bearer token (when OAUTH2_ENABLED=true)."
      }
    }
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}